LayoutTests/imported/w3c/web-platform-tests/screen-orientation/onchange-event-subframe.html

<!DOCTYPE html>
<meta charset="utf-8" />
<meta viewport="width=device-width, initial-scale=1" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script type="module">
  import {
    attachIframe,
    makeCleanup,
    getOppositeOrientation,
  } from "./resources/orientation-utils.js";

  test(() => {
    const iframe = document.createElement("iframe");
    assert_true(
      iframe.sandbox.supports("allow-orientation-lock"),
      "iframe supports 'allow-orientation-lock'"
    );
  });

  promise_test(async (t) => {
    const iframe = await attachIframe({
      src: `./resources/iframe-listen-orientation-change.html`,
      sandbox: "allow-same-origin allow-scripts allow-orientation-lock",
    });
    t.add_cleanup(makeCleanup(iframe));

    let orientations = ["portrait", "landscape"];
    if (screen.orientation.type.includes("portrait")) {
      orientations = orientations.reverse();
    }
    const messageWatcher = new EventWatcher(t, window, "message");
    const changeWatcher = new EventWatcher(t, screen.orientation, "change");

    await test_driver.bless("request fullscreen", null, iframe.contentWindow);
    await iframe.contentDocument.documentElement.requestFullscreen();

    for (const orientation of orientations) {
      const messagePromise = messageWatcher.wait_for("message");
      const eventPromise = changeWatcher.wait_for("change");
      await screen.orientation.lock(orientation);
      const [changeEvent, messageEvent] = await Promise.all([
        eventPromise,
        messagePromise,
      ]);
      assert_true(changeEvent instanceof window.Event, "change event fires");
      assert_equals(changeEvent.type, "change");
      assert_true(
        messageEvent.data.startsWith(orientation),
        `subframe receives orientation change event (got ${orientation})`
      );
    }
  }, "Test cross-origin sub-frame receive orientation change events");

  promise_test(async (t) => {
    const iframe = await attachIframe();
    t.add_cleanup(makeCleanup(iframe));

    // Events
    const topEventPromise = new Promise((r) =>
      screen.orientation.addEventListener("change", r)
    );
    const iframeEventPromise = new Promise((r) =>
      iframe.contentWindow.screen.orientation.addEventListener("change", r)
    );

    // Lock from the iframe
    await test_driver.bless("request fullscreen", null, iframe.contentWindow);
    await iframe.contentDocument.documentElement.requestFullscreen();
    await iframe.contentWindow.screen.orientation.lock(
      getOppositeOrientation()
    );

    const winningEvent = await Promise.race([
      topEventPromise,
      iframeEventPromise,
    ]);
    assert_true(
      winningEvent instanceof window.Event,
      "top-level change event must be fired first"
    );
    const iframeEvent = await iframeEventPromise;
    assert_true(
      iframeEvent instanceof iframe.contentWindow.Event,
      "iframe event eventually fires"
    );
  }, "Check directly that events are fired in right order (from top to bottom)");
</script>
